@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>WinExe</OutputType>
-		<TargetFramework>net5.0-windows</TargetFramework>
+		<TargetFramework>net6.0-windows</TargetFramework>
 		<UseWPF>true</UseWPF>
 		<RootNamespace>FileSearch</RootNamespace>
 		<AssemblyName>FileSearch</AssemblyName>
@@ -11,7 +11,7 @@
 	</PropertyGroup>
 
 	<Target Name="PostPublish" AfterTargets="Publish">
-		<Exec Command="..\UpdateVersion\bin\Release\net5.0\UpdateVersion" />
+		<Exec Command="..\UpdateVersion\bin\Release\net6.0\UpdateVersion" />
 	</Target>
 
 	<ItemGroup>
