@@ -1,427 +1,424 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.IO;
 using System.Text.RegularExpressions;
 
-namespace FileSearch
+namespace FileSearch;
+
+class BackgroundSearch
 {
-	class BackgroundSearch
+
+	#region Members
+
+	SearchInstance searchInstance;
+
+	DateTime lastStatusUpdateTime = DateTime.UtcNow;
+	DateTime startTime = new DateTime();
+	DateTime endTime = new DateTime();
+
+	List<TextAttribute> searchPhrases = new List<TextAttribute>();
+	List<TextAttribute> searchDirectories = new List<TextAttribute>();
+	List<TextAttribute> searchFiles = new List<TextAttribute>();
+
+	List<string> uppercaseIgnoreDirecrories = new List<string>();
+	List<string> uppercaseIgnoreFiles = new List<string>();
+
+	bool regexSearch;
+
+	string currentRoot;
+
+	List<FileHit> searchResults = new List<FileHit>();
+	List<string> searchErrors = new List<string>();
+	List<string> searchIgnores = new List<string>();
+
+	BackgroundWorker backgroundWorker = new BackgroundWorker();
+
+	bool abortPosted = false;
+
+	int searchedFilesCount = 0;
+
+	#endregion
+
+	#region Constructor
+
+	public BackgroundSearch(SearchInstance searchInstance)
 	{
+		this.searchInstance = searchInstance;
 
-		#region Members
+		backgroundWorker.DoWork += BackgroundWorker_DoWork;
+		backgroundWorker.RunWorkerCompleted += BackgroundWorker_RunWorkerCompleted;
 
-		SearchInstance searchInstance;
+		regexSearch = searchInstance.RegexSearch;
 
-		DateTime lastStatusUpdateTime = DateTime.UtcNow;
-		DateTime startTime = new DateTime();
-		DateTime endTime = new DateTime();
-
-		List<TextAttribute> searchPhrases = new List<TextAttribute>();
-		List<TextAttribute> searchDirectories = new List<TextAttribute>();
-		List<TextAttribute> searchFiles = new List<TextAttribute>();
-
-		List<string> uppercaseIgnoreDirecrories = new List<string>();
-		List<string> uppercaseIgnoreFiles = new List<string>();
-
-		bool regexSearch;
-
-		string currentRoot;
-
-		List<FileHit> searchResults = new List<FileHit>();
-		List<string> searchErrors = new List<string>();
-		List<string> searchIgnores = new List<string>();
-
-		BackgroundWorker backgroundWorker = new BackgroundWorker();
-
-		bool abortPosted = false;
-
-		int searchedFilesCount = 0;
-
-		#endregion
-
-		#region Constructor
-
-		public BackgroundSearch(SearchInstance searchInstance)
+		foreach (TextAttribute attribute in searchInstance.SearchPhrases)
 		{
-			this.searchInstance = searchInstance;
-
-			backgroundWorker.DoWork += BackgroundWorker_DoWork;
-			backgroundWorker.RunWorkerCompleted += BackgroundWorker_RunWorkerCompleted;
-
-			regexSearch = searchInstance.RegexSearch;
-
-			foreach (TextAttribute attribute in searchInstance.SearchPhrases)
+			if (!string.IsNullOrEmpty(attribute.Text) && attribute.Used)
 			{
-				if (!string.IsNullOrEmpty(attribute.Text) && attribute.Used)
-				{
-					searchPhrases.Add(new TextAttribute(attribute.Text));
-				}
+				searchPhrases.Add(new TextAttribute(attribute.Text));
 			}
-
-			foreach (TextAttribute attribute in searchInstance.SearchDirectories)
-			{
-				if (!string.IsNullOrEmpty(attribute.Text) && attribute.Used)
-				{
-					searchDirectories.Add(new TextAttribute(attribute.Text));
-				}
-			}
-
-			foreach (TextAttribute attribute in searchInstance.SearchFiles)
-			{
-				if (!string.IsNullOrEmpty(attribute.Text) && attribute.Used)
-				{
-					searchFiles.Add(new TextAttribute(attribute.Text));
-				}
-			}
-			if (searchFiles.Count == 0)
-			{
-				searchFiles.Add(new TextAttribute("*"));
-			}
-
-			foreach (TextAttribute attribute in AppSettings.IgnoredFiles)
-			{
-				uppercaseIgnoreFiles.Add(attribute.UppercaseText);
-			}
-
-			foreach (TextAttribute attribute in AppSettings.IgnoredDirectories)
-			{
-				uppercaseIgnoreDirecrories.Add(attribute.UppercaseText);
-			}
-
-			Search();
-
 		}
 
-		#endregion
-
-		#region Properties
-
-		public bool SearchInProgress
+		foreach (TextAttribute attribute in searchInstance.SearchDirectories)
 		{
-			get { return backgroundWorker.IsBusy; }
-		}
-
-		#endregion
-
-		#region Methods
-
-		private void Search()
-		{
-			backgroundWorker.RunWorkerAsync();
-		}
-
-		public void CancelSearch()
-		{
-			abortPosted = true;
-		}
-
-		private void FindFiles(string searchDirectory)
-		{
-			if (!abortPosted)
+			if (!string.IsNullOrEmpty(attribute.Text) && attribute.Used)
 			{
-				IntPtr INVALID_HANDLE_VALUE = new IntPtr(-1);
-				IntPtr findHandle = WinApi.FindFirstFile(Path.Combine(searchDirectory, "*"), out WIN32_FIND_DATA findData);
+				searchDirectories.Add(new TextAttribute(attribute.Text));
+			}
+		}
 
-				string uppercaseFileName;
-				string newPath;
+		foreach (TextAttribute attribute in searchInstance.SearchFiles)
+		{
+			if (!string.IsNullOrEmpty(attribute.Text) && attribute.Used)
+			{
+				searchFiles.Add(new TextAttribute(attribute.Text));
+			}
+		}
+		if (searchFiles.Count == 0)
+		{
+			searchFiles.Add(new TextAttribute("*"));
+		}
 
-				if (findHandle != INVALID_HANDLE_VALUE)
+		foreach (TextAttribute attribute in AppSettings.IgnoredFiles)
+		{
+			uppercaseIgnoreFiles.Add(attribute.UppercaseText);
+		}
+
+		foreach (TextAttribute attribute in AppSettings.IgnoredDirectories)
+		{
+			uppercaseIgnoreDirecrories.Add(attribute.UppercaseText);
+		}
+
+		Search();
+
+	}
+
+	#endregion
+
+	#region Properties
+
+	public bool SearchInProgress
+	{
+		get { return backgroundWorker.IsBusy; }
+	}
+
+	#endregion
+
+	#region Methods
+
+	private void Search()
+	{
+		backgroundWorker.RunWorkerAsync();
+	}
+
+	public void CancelSearch()
+	{
+		abortPosted = true;
+	}
+
+	private void FindFiles(string searchDirectory)
+	{
+		if (!abortPosted)
+		{
+			IntPtr INVALID_HANDLE_VALUE = new IntPtr(-1);
+			IntPtr findHandle = WinApi.FindFirstFile(Path.Combine(searchDirectory, "*"), out WIN32_FIND_DATA findData);
+
+			string uppercaseFileName;
+			string newPath;
+
+			if (findHandle != INVALID_HANDLE_VALUE)
+			{
+				do
 				{
-					do
+					uppercaseFileName = findData.cFileName.ToUpper();
+					newPath = Path.Combine(searchDirectory, findData.cFileName);
+
+					// Directory
+					if ((findData.dwFileAttributes & FileAttributes.Directory) != 0)
 					{
-						uppercaseFileName = findData.cFileName.ToUpper();
-						newPath = Path.Combine(searchDirectory, findData.cFileName);
-
-						// Directory
-						if ((findData.dwFileAttributes & FileAttributes.Directory) != 0)
+						if (findData.cFileName != "." && findData.cFileName != "..")
 						{
-							if (findData.cFileName != "." && findData.cFileName != "..")
+							if (!DirectoryIsIgnored(uppercaseFileName))
 							{
-								if (!DirectoryIsIgnored(uppercaseFileName))
+								if (searchPhrases.Count == 0 && FileIsMatch(uppercaseFileName))
 								{
-									if (searchPhrases.Count == 0 && FileIsMatch(uppercaseFileName))
-									{
-										searchResults.Add(new FileHit(newPath, searchPhrases, findData, true));
-									}
-									UpdateStatus(newPath);
-									FindFiles(newPath);
+									searchResults.Add(new FileHit(newPath, searchPhrases, findData, true));
 								}
+								UpdateStatus(newPath);
+								FindFiles(newPath);
 							}
 						}
+					}
 
-						// File
-						else
+					// File
+					else
+					{
+						if (FileIsMatch(uppercaseFileName))
 						{
-							if (FileIsMatch(uppercaseFileName))
+							if (!FileIsIgnored(uppercaseFileName))
 							{
-								if (!FileIsIgnored(uppercaseFileName))
+								if (searchPhrases.Count == 0)
 								{
-									if (searchPhrases.Count == 0)
-									{
-										searchResults.Add(new FileHit(newPath, searchPhrases, findData));
-									}
-									else
-									{
-										SearchInFile(newPath, findData);
-									}
-									UpdateStatus(newPath);
+									searchResults.Add(new FileHit(newPath, searchPhrases, findData));
 								}
 								else
 								{
-									searchIgnores.Add(newPath);
+									SearchInFile(newPath, findData);
 								}
+								UpdateStatus(newPath);
 							}
-						}
-					}
-					while (WinApi.FindNextFile(findHandle, out findData) && !abortPosted);
-				}
-
-				WinApi.FindClose(findHandle);
-			}
-		}
-
-		private bool FileIsMatch(string fileName)
-		{
-			foreach (TextAttribute s in searchFiles)
-			{
-				if (WildcardCompare(fileName, s.UppercaseText, false))
-				{
-					return true;
-				}
-			}
-			return false;
-		}
-
-		private bool FileIsIgnored(string fileName)
-		{
-			foreach (string s in uppercaseIgnoreFiles)
-			{
-				if (WildcardCompare(fileName, s, false))
-				{
-					return true;
-				}
-			}
-			return false;
-		}
-
-		private bool DirectoryIsIgnored(string directory)
-		{
-			foreach (string s in uppercaseIgnoreDirecrories)
-			{
-				if (WildcardCompare(directory, s, false))
-				{
-					return true;
-				}
-			}
-			return false;
-		}
-
-		private void UpdateStatus(string currentPath, bool finalUpdate = false)
-		{
-			if (finalUpdate || (DateTime.UtcNow - lastStatusUpdateTime).TotalMilliseconds >= 100)
-			{
-				const string alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ_";
-				int percentageComplete;
-
-				string status;
-				if (currentPath != null)
-				{
-					char firstLetter = Char.ToUpper(currentPath[currentRoot.Length]);
-					int index = alphabet.IndexOf(firstLetter);
-					percentageComplete = index == -1 ? index : (int)((float)(index / (float)alphabet.Length) * 100.0);
-
-					status = currentPath;
-				}
-				else
-				{
-					percentageComplete = 0;
-
-					status = TimeSpanToShortString(endTime.Subtract(startTime));
-				}
-
-				searchInstance.mainWindow.Dispatcher.BeginInvoke(searchInstance.searchProgressUpdateDelegate, new Object[] { searchResults, searchErrors, searchIgnores, status, percentageComplete, searchedFilesCount });
-				lastStatusUpdateTime = DateTime.UtcNow;
-			}
-		}
-
-		private void SearchInFile(string path, WIN32_FIND_DATA findData)
-		{
-			try
-			{
-				FileHit currentHit = null;
-
-				if (regexSearch)
-				{
-					string allText = File.ReadAllText(path);
-					foreach (TextAttribute searchPhrase in searchPhrases)
-					{
-						Match match = Regex.Match(allText, searchPhrase.Text, RegexOptions.Multiline | RegexOptions.IgnoreCase);
-						while (match.Success)
-						{
-							if (currentHit == null)
+							else
 							{
-								currentHit = new FileHit(path, searchPhrases, findData);
+								searchIgnores.Add(newPath);
 							}
-							Match caseSensitiveMatch = Regex.Match(allText[match.Index..], searchPhrase.Text);
-							currentHit.AddPhraseHit(searchPhrase.Text, caseSensitiveMatch.Success && caseSensitiveMatch.Index == 0);
-							match = match.NextMatch();
 						}
 					}
 				}
-
-				else
-				{
-					string allText = File.ReadAllText(path);
-
-					foreach (TextAttribute searchPhrase in searchPhrases)
-					{
-						int i = 0;
-						do
-						{
-							i = allText.IndexOf(searchPhrase.Text, i, StringComparison.OrdinalIgnoreCase);
-							if (i == -1)
-							{
-								break;
-							}
-							if (currentHit == null)
-							{
-								currentHit = new FileHit(path, searchPhrases, findData);
-							}
-							currentHit.AddPhraseHit(searchPhrase.Text, allText.IndexOf(searchPhrase.Text, i, StringComparison.Ordinal) == i);
-
-							i += searchPhrase.Text.Length;
-						}
-						while (true);
-					}
-				}
-
-				if (currentHit != null)
-				{
-					searchResults.Add(currentHit);
-				}
-				searchedFilesCount++;
+				while (WinApi.FindNextFile(findHandle, out findData) && !abortPosted);
 			}
-			catch (Exception e)
+
+			WinApi.FindClose(findHandle);
+		}
+	}
+
+	private bool FileIsMatch(string fileName)
+	{
+		foreach (TextAttribute s in searchFiles)
+		{
+			if (WildcardCompare(fileName, s.UppercaseText, false))
 			{
-				searchErrors.Add(e.Message);
+				return true;
 			}
 		}
+		return false;
+	}
 
-		private bool WildcardCompare(string compare, string wildString, bool ignoreCase)
+	private bool FileIsIgnored(string fileName)
+	{
+		foreach (string s in uppercaseIgnoreFiles)
 		{
-			if (ignoreCase)
+			if (WildcardCompare(fileName, s, false))
 			{
-				wildString = wildString.ToUpper();
-				compare = compare.ToUpper();
+				return true;
+			}
+		}
+		return false;
+	}
+
+	private bool DirectoryIsIgnored(string directory)
+	{
+		foreach (string s in uppercaseIgnoreDirecrories)
+		{
+			if (WildcardCompare(directory, s, false))
+			{
+				return true;
+			}
+		}
+		return false;
+	}
+
+	private void UpdateStatus(string currentPath, bool finalUpdate = false)
+	{
+		if (finalUpdate || (DateTime.UtcNow - lastStatusUpdateTime).TotalMilliseconds >= 100)
+		{
+			const string alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ_";
+			int percentageComplete;
+
+			string status;
+			if (currentPath != null)
+			{
+				char firstLetter = Char.ToUpper(currentPath[currentRoot.Length]);
+				int index = alphabet.IndexOf(firstLetter);
+				percentageComplete = index == -1 ? index : (int)((float)(index / (float)alphabet.Length) * 100.0);
+
+				status = currentPath;
+			}
+			else
+			{
+				percentageComplete = 0;
+
+				status = TimeSpanToShortString(endTime.Subtract(startTime));
 			}
 
-			int wildStringLength = wildString.Length;
-			int CompareLength = compare.Length;
+			searchInstance.mainWindow.Dispatcher.BeginInvoke(searchInstance.searchProgressUpdateDelegate, new Object[] { searchResults, searchErrors, searchIgnores, status, percentageComplete, searchedFilesCount });
+			lastStatusUpdateTime = DateTime.UtcNow;
+		}
+	}
 
-			int wildMatched = wildStringLength;
-			int compareBase = CompareLength;
+	private void SearchInFile(string path, WIN32_FIND_DATA findData)
+	{
+		try
+		{
+			FileHit currentHit = null;
 
-			int wildPosition = 0;
-			int comparePosition = 0;
-
-			// Match until first wildcard '*'
-			while (comparePosition < CompareLength && (wildPosition >= wildStringLength || wildString[wildPosition] != '*'))
+			if (regexSearch)
 			{
-				if (wildPosition >= wildStringLength || (wildString[wildPosition] != compare[comparePosition] && wildString[wildPosition] != '?'))
+				string allText = File.ReadAllText(path);
+				foreach (TextAttribute searchPhrase in searchPhrases)
 				{
-					return false;
-				}
-
-				wildPosition++;
-				comparePosition++;
-			}
-
-			// Process wildcard
-			while (comparePosition < CompareLength)
-			{
-				if (wildPosition < wildStringLength)
-				{
-					if (wildString[wildPosition] == '*')
+					Match match = Regex.Match(allText, searchPhrase.Text, RegexOptions.Multiline | RegexOptions.IgnoreCase);
+					while (match.Success)
 					{
-						wildPosition++;
-
-						if (wildPosition == wildStringLength)
+						if (currentHit == null)
 						{
-							return true;
+							currentHit = new FileHit(path, searchPhrases, findData);
 						}
-
-						wildMatched = wildPosition;
-						compareBase = comparePosition + 1;
-
-						continue;
-					}
-
-					if (wildString[wildPosition] == compare[comparePosition] || wildString[wildPosition] == '?')
-					{
-						wildPosition++;
-						comparePosition++;
-
-						continue;
+						Match caseSensitiveMatch = Regex.Match(allText[match.Index..], searchPhrase.Text);
+						currentHit.AddPhraseHit(searchPhrase.Text, caseSensitiveMatch.Success && caseSensitiveMatch.Index == 0);
+						match = match.NextMatch();
 					}
 				}
-
-				wildPosition = wildMatched;
-				comparePosition = compareBase++;
 			}
 
-			while (wildPosition < wildStringLength && wildString[wildPosition] == '*')
+			else
 			{
-				wildPosition++;
+				string allText = File.ReadAllText(path);
+
+				foreach (TextAttribute searchPhrase in searchPhrases)
+				{
+					int i = 0;
+					do
+					{
+						i = allText.IndexOf(searchPhrase.Text, i, StringComparison.OrdinalIgnoreCase);
+						if (i == -1)
+						{
+							break;
+						}
+						if (currentHit == null)
+						{
+							currentHit = new FileHit(path, searchPhrases, findData);
+						}
+						currentHit.AddPhraseHit(searchPhrase.Text, allText.IndexOf(searchPhrase.Text, i, StringComparison.Ordinal) == i);
+
+						i += searchPhrase.Text.Length;
+					}
+					while (true);
+				}
 			}
 
-			if (wildPosition < wildStringLength)
+			if (currentHit != null)
+			{
+				searchResults.Add(currentHit);
+			}
+			searchedFilesCount++;
+		}
+		catch (Exception e)
+		{
+			searchErrors.Add(e.Message);
+		}
+	}
+
+	private bool WildcardCompare(string compare, string wildString, bool ignoreCase)
+	{
+		if (ignoreCase)
+		{
+			wildString = wildString.ToUpper();
+			compare = compare.ToUpper();
+		}
+
+		int wildStringLength = wildString.Length;
+		int CompareLength = compare.Length;
+
+		int wildMatched = wildStringLength;
+		int compareBase = CompareLength;
+
+		int wildPosition = 0;
+		int comparePosition = 0;
+
+		// Match until first wildcard '*'
+		while (comparePosition < CompareLength && (wildPosition >= wildStringLength || wildString[wildPosition] != '*'))
+		{
+			if (wildPosition >= wildStringLength || (wildString[wildPosition] != compare[comparePosition] && wildString[wildPosition] != '?'))
 			{
 				return false;
 			}
 
-			return true;
+			wildPosition++;
+			comparePosition++;
 		}
 
-		private string TimeSpanToShortString(TimeSpan timeSpan)
+		// Process wildcard
+		while (comparePosition < CompareLength)
 		{
-			if (timeSpan.TotalHours >= 1)
+			if (wildPosition < wildStringLength)
 			{
-				return $"{(int)timeSpan.TotalHours}h {timeSpan.Minutes}m";
+				if (wildString[wildPosition] == '*')
+				{
+					wildPosition++;
+
+					if (wildPosition == wildStringLength)
+					{
+						return true;
+					}
+
+					wildMatched = wildPosition;
+					compareBase = comparePosition + 1;
+
+					continue;
+				}
+
+				if (wildString[wildPosition] == compare[comparePosition] || wildString[wildPosition] == '?')
+				{
+					wildPosition++;
+					comparePosition++;
+
+					continue;
+				}
 			}
-			if (timeSpan.Minutes > 0)
-			{
-				return $"{timeSpan.Minutes}m {timeSpan.Seconds}s";
-			}
-			return $"{timeSpan.Seconds}.{timeSpan.Milliseconds.ToString().PadLeft(3, '0')}s";
+
+			wildPosition = wildMatched;
+			comparePosition = compareBase++;
 		}
 
-		#endregion
-
-		#region Events
-
-		void BackgroundWorker_DoWork(object sender, DoWorkEventArgs e)
+		while (wildPosition < wildStringLength && wildString[wildPosition] == '*')
 		{
-			startTime = DateTime.UtcNow;
-
-			searchInstance.SearchInProgress = true;
-
-			foreach (TextAttribute s in searchDirectories)
-			{
-				currentRoot = s.Text + (s.Text.EndsWith("\\") ? "" : "\\");
-				FindFiles(currentRoot);
-			}
-
-			endTime = DateTime.UtcNow;
+			wildPosition++;
 		}
 
-		void BackgroundWorker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
+		if (wildPosition < wildStringLength)
 		{
-			searchInstance.SearchInProgress = false;
-			UpdateStatus(null, true);
+			return false;
 		}
 
-		#endregion
-
+		return true;
 	}
+
+	private string TimeSpanToShortString(TimeSpan timeSpan)
+	{
+		if (timeSpan.TotalHours >= 1)
+		{
+			return $"{(int)timeSpan.TotalHours}h {timeSpan.Minutes}m";
+		}
+		if (timeSpan.Minutes > 0)
+		{
+			return $"{timeSpan.Minutes}m {timeSpan.Seconds}s";
+		}
+		return $"{timeSpan.Seconds}.{timeSpan.Milliseconds.ToString().PadLeft(3, '0')}s";
+	}
+
+	#endregion
+
+	#region Events
+
+	void BackgroundWorker_DoWork(object sender, DoWorkEventArgs e)
+	{
+		startTime = DateTime.UtcNow;
+
+		searchInstance.SearchInProgress = true;
+
+		foreach (TextAttribute s in searchDirectories)
+		{
+			currentRoot = s.Text + (s.Text.EndsWith("\\") ? "" : "\\");
+			FindFiles(currentRoot);
+		}
+
+		endTime = DateTime.UtcNow;
+	}
+
+	void BackgroundWorker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
+	{
+		searchInstance.SearchInProgress = false;
+		UpdateStatus(null, true);
+	}
+
+	#endregion
+
 }
