@@ -1,27 +1,24 @@
-﻿namespace FileSearch
+﻿namespace FileSearch;
+
+public enum TextState
 {
+	Miss,
+	Hit,
+	Header,
+	Filler,
+	Surround,
+	SurroundSpacing
+}
 
-	public enum TextState
-	{
-		Miss,
-		Hit,
-		Header,
-		Filler,
-		Surround,
-		SurroundSpacing
-	}
+public enum NewlineMode
+{
+	Windows,
+	Unix,
+	Mac
+}
 
-	public enum NewlineMode
-	{
-		Windows,
-		Unix,
-		Mac
-	}
-
-	public enum LogWindowType
-	{
-		Errors,
-		IgnoredFiles
-	}
-
+public enum LogWindowType
+{
+	Errors,
+	IgnoredFiles
 }
