@@ -1,24 +1,23 @@
-﻿namespace FileSearch
+﻿namespace FileSearch;
+
+class RegexHit
 {
-	class RegexHit
+	public RegexHit()
 	{
-		public RegexHit()
-		{
-		}
+	}
 
-		public RegexHit(int start, int length)
-		{
-			Start = start;
-			Length = length;
-		}
+	public RegexHit(int start, int length)
+	{
+		Start = start;
+		Length = length;
+	}
 
-		public int Start { get; set; }
+	public int Start { get; set; }
 
-		public int Length { get; set; }
+	public int Length { get; set; }
 
-		public override string ToString()
-		{
-			return Start + " " + Length;
-		}
+	public override string ToString()
+	{
+		return Start + " " + Length;
 	}
 }
