@@ -1,43 +1,42 @@
 ﻿using System.Windows;
 
-namespace FileSearch
+namespace FileSearch;
+
+public partial class RenameTabWindow : Window
 {
-	public partial class RenameTabWindow : Window
+
+	#region Constructor
+
+	public RenameTabWindow()
 	{
-
-		#region Constructor
-
-		public RenameTabWindow()
-		{
-			InitializeComponent();
-		}
-
-		#endregion
-
-		#region Properties
-
-		public string TabName
-		{
-			get { return TextBoxName.Text; }
-			set { TextBoxName.Text = value; }
-		}
-
-		#endregion
-
-		#region Events
-
-		private void ButtonOk_Click(object sender, RoutedEventArgs e)
-		{
-			DialogResult = true;
-		}
-
-		private void Window_Loaded(object sender, RoutedEventArgs e)
-		{
-			TextBoxName.Focus();
-			TextBoxName.SelectAll();
-		}
-
-		#endregion
-
+		InitializeComponent();
 	}
+
+	#endregion
+
+	#region Properties
+
+	public string TabName
+	{
+		get { return TextBoxName.Text; }
+		set { TextBoxName.Text = value; }
+	}
+
+	#endregion
+
+	#region Events
+
+	private void ButtonOk_Click(object sender, RoutedEventArgs e)
+	{
+		DialogResult = true;
+	}
+
+	private void Window_Loaded(object sender, RoutedEventArgs e)
+	{
+		TextBoxName.Focus();
+		TextBoxName.SelectAll();
+	}
+
+	#endregion
+
 }
